+++ src/shared/storeKey.ts
@@ -33,3 +33,7 @@ export const LOCALE = key<Language>('LOCALE')
 export const DANMAKU_VISIBLE = key<boolean>('DANMAKU_VISIBLE')
 
 export const NEED_RELOAD = key<boolean>('NEED_RELOAD')
+
+export const NATIVE_HOST_MISSING_REMINDER_DISABLED = key<boolean>(
+  'NATIVE_HOST_MISSING_REMINDER_DISABLED',
+)